requirements.txt
fastapi[all]
Pillow
python-multipart
gunicorn==20.1.0
